migrate strike website to new framework code
